refactor: make use of 'isNotNullOrUndefined' and add tslint rule to enforce file name casing
