--- src/util/is-not-null-or-undefined.ts
@@ -1,1 +1,0 @@
-export const isNotNullOrUndefined = <T>(input: null | undefined | T): input is T => input !== null && input !== undefined;
